@@ -15,15 +15,16 @@
     }
   ],
   "require": {
-    "php": ">=5.3.3",
+    "php": ">=5.4.0",
     "ext-curl": "*",
     "ext-json": "*",
     "ext-mbstring": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "~4.0",
-    "satooshi/php-coveralls": "~0.6.1",
-    "squizlabs/php_codesniffer": "~2.0"
+    "php-coveralls/php-coveralls": "1.*",
+    "squizlabs/php_codesniffer": "~2.0",
+    "symfony/process": "~2.8"
   },
   "autoload": {
     "psr-4": { "Stripe\\" : "lib/" }
