@@ -6,13 +6,23 @@ use Stripe\Error;
 
 class RequestOptions
 {
+    /**
+     * @var array A list of headers that should be persisted across requests.
+     */
+    public static $HEADERS_TO_PERSIST = [
+        'Stripe-Account',
+        'Stripe-Version',
+    ];
+
     public $headers;
     public $apiKey;
+    public $apiBase;
 
-    public function __construct($key = null, $headers = array())
+    public function __construct($key = null, $headers = [], $base = null)
     {
         $this->apiKey = $key;
         $this->headers = $headers;
+        $this->apiBase = $base;
     }
 
     /**
@@ -28,8 +38,23 @@ class RequestOptions
         if ($other_options->apiKey === null) {
             $other_options->apiKey = $this->apiKey;
         }
+        if ($other_options->apiBase === null) {
+            $other_options->apiBase = $this->apiBase;
+        }
         $other_options->headers = array_merge($this->headers, $other_options->headers);
         return $other_options;
+    }
+
+    /**
+     * Discards all headers that we don't want to persist across requests.
+     */
+    public function discardNonPersistentHeaders()
+    {
+        foreach ($this->headers as $k => $v) {
+            if (!in_array($k, self::$HEADERS_TO_PERSIST)) {
+                unset($this->headers[$k]);
+            }
+        }
     }
 
     /**
@@ -45,16 +70,17 @@ class RequestOptions
         }
 
         if (is_null($options)) {
-            return new RequestOptions(null, array());
+            return new RequestOptions(null, [], null);
         }
 
         if (is_string($options)) {
-            return new RequestOptions($options, array());
+            return new RequestOptions($options, [], null);
         }
 
         if (is_array($options)) {
-            $headers = array();
+            $headers = [];
             $key = null;
+            $base = null;
             if (array_key_exists('api_key', $options)) {
                 $key = $options['api_key'];
             }
@@ -67,7 +93,10 @@ class RequestOptions
             if (array_key_exists('stripe_version', $options)) {
                 $headers['Stripe-Version'] = $options['stripe_version'];
             }
-            return new RequestOptions($key, $headers);
+            if (array_key_exists('api_base', $options)) {
+                $base = $options['api_base'];
+            }
+            return new RequestOptions($key, $headers, $base);
         }
 
         $message = 'The second argument to Stripe API method calls is an '
