@@ -9,9 +9,9 @@ class Set implements IteratorAggregate
 {
     private $_elts;
 
-    public function __construct($members = array())
+    public function __construct($members = [])
     {
-        $this->_elts = array();
+        $this->_elts = [];
         foreach ($members as $item) {
             $this->_elts[$item] = true;
         }
