@@ -11,21 +11,22 @@ abstract class Util
 
     /**
      * Whether the provided array (or other) is a list rather than a dictionary.
+     * A list is defined as an array for which all the keys are consecutive
+     * integers starting at 0. Empty arrays are considered to be lists.
      *
      * @param array|mixed $array
-     * @return boolean True if the given object is a list.
+     * @return boolean true if the given object is a list.
      */
     public static function isList($array)
     {
         if (!is_array($array)) {
             return false;
         }
-
-      // TODO: generally incorrect, but it's correct given Stripe's response
-        foreach (array_keys($array) as $k) {
-            if (!is_numeric($k)) {
-                return false;
-            }
+        if ($array === []) {
+            return true;
+        }
+        if (array_keys($array) !== range(0, count($array) - 1)) {
+            return false;
         }
         return true;
     }
@@ -38,7 +39,7 @@ abstract class Util
      */
     public static function convertStripeObjectToArray($values)
     {
-        $results = array();
+        $results = [];
         foreach ($values as $k => $v) {
             // FIXME: this is an encapsulation violation
             if ($k[0] == '_') {
@@ -64,47 +65,91 @@ abstract class Util
      */
     public static function convertToStripeObject($resp, $opts)
     {
-        $types = array(
-            'account' => 'Stripe\\Account',
-            'alipay_account' => 'Stripe\\AlipayAccount',
-            'apple_pay_domain' => 'Stripe\\ApplePayDomain',
-            'bank_account' => 'Stripe\\BankAccount',
-            'balance_transaction' => 'Stripe\\BalanceTransaction',
-            'card' => 'Stripe\\Card',
-            'charge' => 'Stripe\\Charge',
-            'country_spec' => 'Stripe\\CountrySpec',
-            'coupon' => 'Stripe\\Coupon',
-            'customer' => 'Stripe\\Customer',
-            'dispute' => 'Stripe\\Dispute',
-            'ephemeral_key' => 'Stripe\\EphemeralKey',
-            'list' => 'Stripe\\Collection',
-            'login_link' => 'Stripe\\LoginLink',
-            'invoice' => 'Stripe\\Invoice',
-            'invoiceitem' => 'Stripe\\InvoiceItem',
-            'event' => 'Stripe\\Event',
-            'file_upload' => 'Stripe\\FileUpload',
-            'token' => 'Stripe\\Token',
-            'transfer' => 'Stripe\\Transfer',
-            'transfer_reversal' => 'Stripe\\TransferReversal',
-            'order' => 'Stripe\\Order',
-            'order_return' => 'Stripe\\OrderReturn',
-            'payout' => 'Stripe\\Payout',
-            'plan' => 'Stripe\\Plan',
-            'product' => 'Stripe\\Product',
-            'recipient' => 'Stripe\\Recipient',
-            'recipient_transfer' => 'Stripe\\RecipientTransfer',
-            'refund' => 'Stripe\\Refund',
-            'sku' => 'Stripe\\SKU',
-            'source' => 'Stripe\\Source',
-            'subscription' => 'Stripe\\Subscription',
-            'subscription_item' => 'Stripe\\SubscriptionItem',
-            'three_d_secure' => 'Stripe\\ThreeDSecure',
-            'fee_refund' => 'Stripe\\ApplicationFeeRefund',
-            'bitcoin_receiver' => 'Stripe\\BitcoinReceiver',
-            'bitcoin_transaction' => 'Stripe\\BitcoinTransaction',
-        );
+        $types = [
+            // data structures
+            \Stripe\Collection::OBJECT_NAME => 'Stripe\\Collection',
+
+            // business objects
+            \Stripe\Account::OBJECT_NAME => 'Stripe\\Account',
+            \Stripe\AccountLink::OBJECT_NAME => 'Stripe\\AccountLink',
+            \Stripe\AlipayAccount::OBJECT_NAME => 'Stripe\\AlipayAccount',
+            \Stripe\ApplePayDomain::OBJECT_NAME => 'Stripe\\ApplePayDomain',
+            \Stripe\ApplicationFee::OBJECT_NAME => 'Stripe\\ApplicationFee',
+            \Stripe\Balance::OBJECT_NAME => 'Stripe\\Balance',
+            \Stripe\BalanceTransaction::OBJECT_NAME => 'Stripe\\BalanceTransaction',
+            \Stripe\BankAccount::OBJECT_NAME => 'Stripe\\BankAccount',
+            \Stripe\BitcoinReceiver::OBJECT_NAME => 'Stripe\\BitcoinReceiver',
+            \Stripe\BitcoinTransaction::OBJECT_NAME => 'Stripe\\BitcoinTransaction',
+            \Stripe\Capability::OBJECT_NAME => 'Stripe\\Capability',
+            \Stripe\Card::OBJECT_NAME => 'Stripe\\Card',
+            \Stripe\Charge::OBJECT_NAME => 'Stripe\\Charge',
+            \Stripe\Checkout\Session::OBJECT_NAME => 'Stripe\\Checkout\\Session',
+            \Stripe\CountrySpec::OBJECT_NAME => 'Stripe\\CountrySpec',
+            \Stripe\Coupon::OBJECT_NAME => 'Stripe\\Coupon',
+            \Stripe\CreditNote::OBJECT_NAME => 'Stripe\\CreditNote',
+            \Stripe\Customer::OBJECT_NAME => 'Stripe\\Customer',
+            \Stripe\Discount::OBJECT_NAME => 'Stripe\\Discount',
+            \Stripe\Dispute::OBJECT_NAME => 'Stripe\\Dispute',
+            \Stripe\EphemeralKey::OBJECT_NAME => 'Stripe\\EphemeralKey',
+            \Stripe\Event::OBJECT_NAME => 'Stripe\\Event',
+            \Stripe\ExchangeRate::OBJECT_NAME => 'Stripe\\ExchangeRate',
+            \Stripe\ApplicationFeeRefund::OBJECT_NAME => 'Stripe\\ApplicationFeeRefund',
+            \Stripe\File::OBJECT_NAME => 'Stripe\\File',
+            \Stripe\File::OBJECT_NAME_ALT => 'Stripe\\File',
+            \Stripe\FileLink::OBJECT_NAME => 'Stripe\\FileLink',
+            \Stripe\Invoice::OBJECT_NAME => 'Stripe\\Invoice',
+            \Stripe\InvoiceItem::OBJECT_NAME => 'Stripe\\InvoiceItem',
+            \Stripe\InvoiceLineItem::OBJECT_NAME => 'Stripe\\InvoiceLineItem',
+            \Stripe\IssuerFraudRecord::OBJECT_NAME => 'Stripe\\IssuerFraudRecord',
+            \Stripe\Issuing\Authorization::OBJECT_NAME => 'Stripe\\Issuing\\Authorization',
+            \Stripe\Issuing\Card::OBJECT_NAME => 'Stripe\\Issuing\\Card',
+            \Stripe\Issuing\CardDetails::OBJECT_NAME => 'Stripe\\Issuing\\CardDetails',
+            \Stripe\Issuing\Cardholder::OBJECT_NAME => 'Stripe\\Issuing\\Cardholder',
+            \Stripe\Issuing\Dispute::OBJECT_NAME => 'Stripe\\Issuing\\Dispute',
+            \Stripe\Issuing\Transaction::OBJECT_NAME => 'Stripe\\Issuing\\Transaction',
+            \Stripe\LoginLink::OBJECT_NAME => 'Stripe\\LoginLink',
+            \Stripe\Order::OBJECT_NAME => 'Stripe\\Order',
+            \Stripe\OrderItem::OBJECT_NAME => 'Stripe\\OrderItem',
+            \Stripe\OrderReturn::OBJECT_NAME => 'Stripe\\OrderReturn',
+            \Stripe\PaymentIntent::OBJECT_NAME => 'Stripe\\PaymentIntent',
+            \Stripe\PaymentMethod::OBJECT_NAME => 'Stripe\\PaymentMethod',
+            \Stripe\Payout::OBJECT_NAME => 'Stripe\\Payout',
+            \Stripe\Person::OBJECT_NAME => 'Stripe\\Person',
+            \Stripe\Plan::OBJECT_NAME => 'Stripe\\Plan',
+            \Stripe\Product::OBJECT_NAME => 'Stripe\\Product',
+            \Stripe\Radar\EarlyFraudWarning::OBJECT_NAME => 'Stripe\\Radar\\EarlyFraudWarning',
+            \Stripe\Radar\ValueList::OBJECT_NAME => 'Stripe\\Radar\\ValueList',
+            \Stripe\Radar\ValueListItem::OBJECT_NAME => 'Stripe\\Radar\\ValueListItem',
+            \Stripe\Recipient::OBJECT_NAME => 'Stripe\\Recipient',
+            \Stripe\RecipientTransfer::OBJECT_NAME => 'Stripe\\RecipientTransfer',
+            \Stripe\Refund::OBJECT_NAME => 'Stripe\\Refund',
+            \Stripe\Reporting\ReportRun::OBJECT_NAME => 'Stripe\\Reporting\\ReportRun',
+            \Stripe\Reporting\ReportType::OBJECT_NAME => 'Stripe\\Reporting\\ReportType',
+            \Stripe\Review::OBJECT_NAME => 'Stripe\\Review',
+            \Stripe\SKU::OBJECT_NAME => 'Stripe\\SKU',
+            \Stripe\Sigma\ScheduledQueryRun::OBJECT_NAME => 'Stripe\\Sigma\\ScheduledQueryRun',
+            \Stripe\Source::OBJECT_NAME => 'Stripe\\Source',
+            \Stripe\SourceTransaction::OBJECT_NAME => 'Stripe\\SourceTransaction',
+            \Stripe\Subscription::OBJECT_NAME => 'Stripe\\Subscription',
+            \Stripe\SubscriptionItem::OBJECT_NAME => 'Stripe\\SubscriptionItem',
+            \Stripe\SubscriptionSchedule::OBJECT_NAME => 'Stripe\\SubscriptionSchedule',
+            \Stripe\SubscriptionScheduleRevision::OBJECT_NAME => 'Stripe\\SubscriptionScheduleRevision',
+            \Stripe\TaxId::OBJECT_NAME => 'Stripe\\TaxId',
+            \Stripe\TaxRate::OBJECT_NAME => 'Stripe\\TaxRate',
+            \Stripe\ThreeDSecure::OBJECT_NAME => 'Stripe\\ThreeDSecure',
+            \Stripe\Terminal\ConnectionToken::OBJECT_NAME => 'Stripe\\Terminal\\ConnectionToken',
+            \Stripe\Terminal\Location::OBJECT_NAME => 'Stripe\\Terminal\\Location',
+            \Stripe\Terminal\Reader::OBJECT_NAME => 'Stripe\\Terminal\\Reader',
+            \Stripe\Token::OBJECT_NAME => 'Stripe\\Token',
+            \Stripe\Topup::OBJECT_NAME => 'Stripe\\Topup',
+            \Stripe\Transfer::OBJECT_NAME => 'Stripe\\Transfer',
+            \Stripe\TransferReversal::OBJECT_NAME => 'Stripe\\TransferReversal',
+            \Stripe\UsageRecord::OBJECT_NAME => 'Stripe\\UsageRecord',
+            \Stripe\UsageRecordSummary::OBJECT_NAME => 'Stripe\\UsageRecordSummary',
+            \Stripe\WebhookEndpoint::OBJECT_NAME => 'Stripe\\WebhookEndpoint',
+        ];
         if (self::isList($resp)) {
-            $mapped = array();
+            $mapped = [];
             foreach ($resp as $i) {
                 array_push($mapped, self::convertToStripeObject($i, $opts));
             }
@@ -177,41 +222,138 @@ abstract class Util
     }
 
     /**
-     * @param array $arr A map of param keys to values.
-     * @param string|null $prefix
+     * Recursively goes through an array of parameters. If a parameter is an instance of
+     * ApiResource, then it is replaced by the resource's ID.
+     * Also clears out null values.
      *
-     * @return string A querystring, essentially.
+     * @param mixed $h
+     * @return mixed
      */
-    public static function urlEncode($arr, $prefix = null)
+    public static function objectsToIds($h)
     {
-        if (!is_array($arr)) {
-            return $arr;
+        if ($h instanceof \Stripe\ApiResource) {
+            return $h->id;
+        } elseif (static::isList($h)) {
+            $results = [];
+            foreach ($h as $v) {
+                array_push($results, static::objectsToIds($v));
+            }
+            return $results;
+        } elseif (is_array($h)) {
+            $results = [];
+            foreach ($h as $k => $v) {
+                if (is_null($v)) {
+                    continue;
+                }
+                $results[$k] = static::objectsToIds($v);
+            }
+            return $results;
+        } else {
+            return $h;
         }
+    }
 
-        $r = array();
-        foreach ($arr as $k => $v) {
-            if (is_null($v)) {
-                continue;
-            }
+    /**
+     * @param array $params
+     *
+     * @return string
+     */
+    public static function encodeParameters($params)
+    {
+        $flattenedParams = self::flattenParams($params);
+        $pieces = [];
+        foreach ($flattenedParams as $param) {
+            list($k, $v) = $param;
+            array_push($pieces, self::urlEncode($k) . '=' . self::urlEncode($v));
+        }
+        return implode('&', $pieces);
+    }
 
-            if ($prefix) {
-                if ($k !== null && (!is_int($k) || is_array($v))) {
-                    $k = $prefix."[".$k."]";
-                } else {
-                    $k = $prefix."[]";
-                }
-            }
+    /**
+     * @param array $params
+     * @param string|null $parentKey
+     *
+     * @return array
+     */
+    public static function flattenParams($params, $parentKey = null)
+    {
+        $result = [];
 
-            if (is_array($v)) {
-                $enc = self::urlEncode($v, $k);
-                if ($enc) {
-                    $r[] = $enc;
-                }
+        foreach ($params as $key => $value) {
+            $calculatedKey = $parentKey ? "{$parentKey}[{$key}]" : $key;
+
+            if (self::isList($value)) {
+                $result = array_merge($result, self::flattenParamsList($value, $calculatedKey));
+            } elseif (is_array($value)) {
+                $result = array_merge($result, self::flattenParams($value, $calculatedKey));
             } else {
-                $r[] = urlencode($k)."=".urlencode($v);
+                array_push($result, [$calculatedKey, $value]);
             }
         }
 
-        return implode("&", $r);
+        return $result;
+    }
+
+    /**
+     * @param array $value
+     * @param string $calculatedKey
+     *
+     * @return array
+     */
+    public static function flattenParamsList($value, $calculatedKey)
+    {
+        $result = [];
+
+        foreach ($value as $i => $elem) {
+            if (self::isList($elem)) {
+                $result = array_merge($result, self::flattenParamsList($elem, $calculatedKey));
+            } elseif (is_array($elem)) {
+                $result = array_merge($result, self::flattenParams($elem, "{$calculatedKey}[{$i}]"));
+            } else {
+                array_push($result, ["{$calculatedKey}[{$i}]", $elem]);
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param string $key A string to URL-encode.
+     *
+     * @return string The URL-encoded string.
+     */
+    public static function urlEncode($key)
+    {
+        $s = urlencode($key);
+
+        // Don't use strict form encoding by changing the square bracket control
+        // characters back to their literals. This is fine by the server, and
+        // makes these parameter strings easier to read.
+        $s = str_replace('%5B', '[', $s);
+        $s = str_replace('%5D', ']', $s);
+
+        return $s;
+    }
+
+    public static function normalizeId($id)
+    {
+        if (is_array($id)) {
+            $params = $id;
+            $id = $params['id'];
+            unset($params['id']);
+        } else {
+            $params = [];
+        }
+        return [$id, $params];
+    }
+
+    /**
+     * Returns UNIX timestamp in milliseconds
+     *
+     * @return integer current time in millis
+     */
+    public static function currentTimeMillis()
+    {
+        return (int) round(microtime(true) * 1000);
     }
 }
