@@ -9,11 +9,42 @@ namespace Stripe;
  */
 abstract class ApiResource extends StripeObject
 {
-    private static $HEADERS_TO_PERSIST = array('Stripe-Account' => true, 'Stripe-Version' => true);
+    use ApiOperations\Request;
 
-    public static function baseUrl()
+    /**
+     * @return \Stripe\Util\Set A list of fields that can be their own type of
+     * API resource (say a nested card under an account for example), and if
+     * that resource is set, it should be transmitted to the API on a create or
+     * update. Doing so is not the default behavior because API resources
+     * should normally be persisted on their own RESTful endpoints.
+     */
+    public static function getSavedNestedResources()
     {
-        return Stripe::$apiBase;
+        static $savedNestedResources = null;
+        if ($savedNestedResources === null) {
+            $savedNestedResources = new Util\Set();
+        }
+        return $savedNestedResources;
+    }
+
+    /**
+     * @var boolean A flag that can be set a behavior that will cause this
+     * resource to be encoded and sent up along with an update of its parent
+     * resource. This is usually not desirable because resources are updated
+     * individually on their own endpoints, but there are certain cases,
+     * replacing a customer's source for example, where this is allowed.
+     */
+    public $saveWithParent = false;
+
+    public function __set($k, $v)
+    {
+        parent::__set($k, $v);
+        $v = $this->$k;
+        if ((static::getSavedNestedResources()->includes($k)) &&
+            ($v instanceof ApiResource)) {
+            $v->saveWithParent = true;
+        }
+        return $v;
     }
 
     /**
@@ -36,27 +67,11 @@ abstract class ApiResource extends StripeObject
     }
 
     /**
-     * @return string The name of the class, with namespacing and underscores
-     *    stripped.
+     * @return string The base URL for the given class.
      */
-    public static function className()
+    public static function baseUrl()
     {
-        $class = get_called_class();
-        // Useful for namespaces: Foo\Charge
-        if ($postfixNamespaces = strrchr($class, '\\')) {
-            $class = substr($postfixNamespaces, 1);
-        }
-        // Useful for underscored 'namespaces': Foo_Charge
-        if ($postfixFakeNamespaces = strrchr($class, '')) {
-            $class = $postfixFakeNamespaces;
-        }
-        if (substr($class, 0, strlen('Stripe')) == 'Stripe') {
-            $class = substr($class, strlen('Stripe'));
-        }
-        $class = str_replace('_', '', $class);
-        $name = urlencode($class);
-        $name = strtolower($name);
-        return $name;
+        return Stripe::$apiBase;
     }
 
     /**
@@ -64,7 +79,9 @@ abstract class ApiResource extends StripeObject
      */
     public static function classUrl()
     {
-        $base = static::className();
+        // Replace dots with slashes for namespaced resources, e.g. if the object's name is
+        // "foo.bar", then its URL will be "/v1/foo/bars".
+        $base = str_replace('.', '/', static::OBJECT_NAME);
         return "/v1/${base}s";
     }
 
@@ -91,112 +108,5 @@ abstract class ApiResource extends StripeObject
     public function instanceUrl()
     {
         return static::resourceUrl($this['id']);
-    }
-
-    protected static function _validateParams($params = null)
-    {
-        if ($params && !is_array($params)) {
-            $message = "You must pass an array as the first argument to Stripe API "
-               . "method calls.  (HINT: an example call to create a charge "
-               . "would be: \"Stripe\\Charge::create(array('amount' => 100, "
-               . "'currency' => 'usd', 'source' => 'tok_1234'))\")";
-            throw new Error\Api($message);
-        }
-    }
-
-    protected function _request($method, $url, $params = array(), $options = null)
-    {
-        $opts = $this->_opts->merge($options);
-        list($resp, $options) = static::_staticRequest($method, $url, $params, $opts);
-        $this->setLastResponse($resp);
-        return array($resp->json, $options);
-    }
-
-    protected static function _staticRequest($method, $url, $params, $options)
-    {
-        $opts = Util\RequestOptions::parse($options);
-        $requestor = new ApiRequestor($opts->apiKey, static::baseUrl());
-        list($response, $opts->apiKey) = $requestor->request($method, $url, $params, $opts->headers);
-        foreach ($opts->headers as $k => $v) {
-            if (!array_key_exists($k, self::$HEADERS_TO_PERSIST)) {
-                unset($opts->headers[$k]);
-            }
-        }
-        return array($response, $opts);
-    }
-
-    protected static function _retrieve($id, $options = null)
-    {
-        $opts = Util\RequestOptions::parse($options);
-        $instance = new static($id, $opts);
-        $instance->refresh();
-        return $instance;
-    }
-
-    protected static function _all($params = null, $options = null)
-    {
-        self::_validateParams($params);
-        $url = static::classUrl();
-
-        list($response, $opts) = static::_staticRequest('get', $url, $params, $options);
-        $obj = Util\Util::convertToStripeObject($response->json, $opts);
-        if (!is_a($obj, 'Stripe\\Collection')) {
-            $class = get_class($obj);
-            $message = "Expected type \"Stripe\\Collection\", got \"$class\" instead";
-            throw new Error\Api($message);
-        }
-        $obj->setLastResponse($response);
-        $obj->setRequestParams($params);
-        return $obj;
-    }
-
-    protected static function _create($params = null, $options = null)
-    {
-        self::_validateParams($params);
-        $url = static::classUrl();
-
-        list($response, $opts) = static::_staticRequest('post', $url, $params, $options);
-        $obj = Util\Util::convertToStripeObject($response->json, $opts);
-        $obj->setLastResponse($response);
-        return $obj;
-    }
-
-    /**
-     * @param string $id The ID of the API resource to update.
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return ApiResource the updated API resource
-     */
-    protected static function _update($id, $params = null, $options = null)
-    {
-        self::_validateParams($params);
-        $url = static::resourceUrl($id);
-
-        list($response, $opts) = static::_staticRequest('post', $url, $params, $options);
-        $obj = Util\Util::convertToStripeObject($response->json, $opts);
-        $obj->setLastResponse($response);
-        return $obj;
-    }
-
-    protected function _save($options = null)
-    {
-        $params = $this->serializeParameters();
-        if (count($params) > 0) {
-            $url = $this->instanceUrl();
-            list($response, $opts) = $this->_request('post', $url, $params, $options);
-            $this->refreshFrom($response, $opts);
-        }
-        return $this;
-    }
-
-    protected function _delete($params = null, $options = null)
-    {
-        self::_validateParams($params);
-
-        $url = $this->instanceUrl();
-        list($response, $opts) = $this->_request('delete', $url, $params, $options);
-        $this->refreshFrom($response, $opts);
-        return $this;
     }
 }
