@@ -6,14 +6,18 @@ namespace Stripe;
  * Class Balance
  *
  * @property string $object
- * @property mixed $available
- * @property bool $livedmode
- * @property mixed $pending
+ * @property array $available
+ * @property array $connect_reserved
+ * @property bool $livemode
+ * @property array $pending
  *
  * @package Stripe
  */
 class Balance extends SingletonApiResource
 {
+
+    const OBJECT_NAME = "balance";
+
     /**
      * @param array|string|null $opts
      *
