@@ -8,90 +8,62 @@ namespace Stripe;
  * @property string $id
  * @property string $object
  * @property int $account_balance
- * @property string $business_vat_id
+ * @property mixed $address
  * @property string $created
  * @property string $currency
  * @property string $default_source
  * @property bool $delinquent
  * @property string $description
- * @property mixed $discount
+ * @property Discount $discount
  * @property string $email
+ * @property string $invoice_prefix
+ * @property mixed $invoice_settings
  * @property bool $livemode
- * @property array $metadata
+ * @property StripeObject $metadata
+ * @property string $name
+ * @property string $phone
+ * @property string[] preferred_locales
  * @property mixed $shipping
  * @property Collection $sources
  * @property Collection $subscriptions
+ * @property string $tax_exempt
+ * @property Collection $tax_ids
  *
  * @package Stripe
  */
 class Customer extends ApiResource
 {
-    /**
-     * @param array|string $id The ID of the customer to retrieve, or an
-     *     options array containing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return Customer
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
+
+    const OBJECT_NAME = "customer";
+
+    use ApiOperations\All;
+    use ApiOperations\Create;
+    use ApiOperations\Delete;
+    use ApiOperations\NestedResource;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
 
     /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of Customers
+     * Possible string representations of the customer's type of tax exemption.
+     * @link https://stripe.com/docs/api/customers/object#customer_object-tax_exempt
      */
-    public static function all($params = null, $opts = null)
+    const TAX_EXEMPT_NONE    = 'none';
+    const TAX_EXEMPT_EXEMPT  = 'exempt';
+    const TAX_EXEMPT_REVERSE = 'reverse';
+
+    public static function getSavedNestedResources()
     {
-        return self::_all($params, $opts);
+        static $savedNestedResources = null;
+        if ($savedNestedResources === null) {
+            $savedNestedResources = new Util\Set([
+                'source',
+            ]);
+        }
+        return $savedNestedResources;
     }
 
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Customer The created customer.
-     */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
-
-    /**
-     * @param string $id The ID of the customer to update.
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Customer The updated customer.
-     */
-    public static function update($id, $params = null, $options = null)
-    {
-        return self::_update($id, $params, $options);
-    }
-
-    /**
-     * @param array|string|null $opts
-     *
-     * @return Customer The saved customer.
-     */
-    public function save($opts = null)
-    {
-        return $this->_save($opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Customer The deleted customer.
-     */
-    public function delete($params = null, $opts = null)
-    {
-        return $this->_delete($params, $opts);
-    }
+    const PATH_SOURCES = '/sources';
+    const PATH_TAX_IDS = '/tax_ids';
 
     /**
      * @param array|null $params
@@ -100,9 +72,7 @@ class Customer extends ApiResource
      */
     public function addInvoiceItem($params = null)
     {
-        if (!$params) {
-            $params = array();
-        }
+        $params = $params ?: [];
         $params['customer'] = $this->id;
         $ii = InvoiceItem::create($params, $this->_opts);
         return $ii;
@@ -115,9 +85,7 @@ class Customer extends ApiResource
      */
     public function invoices($params = null)
     {
-        if (!$params) {
-            $params = array();
-        }
+        $params = $params ?: [];
         $params['customer'] = $this->id;
         $invoices = Invoice::all($params, $this->_opts);
         return $invoices;
@@ -130,9 +98,7 @@ class Customer extends ApiResource
      */
     public function invoiceItems($params = null)
     {
-        if (!$params) {
-            $params = array();
-        }
+        $params = $params ?: [];
         $params['customer'] = $this->id;
         $iis = InvoiceItem::all($params, $this->_opts);
         return $iis;
@@ -145,9 +111,7 @@ class Customer extends ApiResource
      */
     public function charges($params = null)
     {
-        if (!$params) {
-            $params = array();
-        }
+        $params = $params ?: [];
         $params['customer'] = $this->id;
         $charges = Charge::all($params, $this->_opts);
         return $charges;
@@ -162,7 +126,7 @@ class Customer extends ApiResource
     {
         $url = $this->instanceUrl() . '/subscription';
         list($response, $opts) = $this->_request('post', $url, $params);
-        $this->refreshFrom(array('subscription' => $response), $opts, true);
+        $this->refreshFrom(['subscription' => $response], $opts, true);
         return $this->subscription;
     }
 
@@ -175,7 +139,7 @@ class Customer extends ApiResource
     {
         $url = $this->instanceUrl() . '/subscription';
         list($response, $opts) = $this->_request('delete', $url, $params);
-        $this->refreshFrom(array('subscription' => $response), $opts, true);
+        $this->refreshFrom(['subscription' => $response], $opts, true);
         return $this->subscription;
     }
 
@@ -186,6 +150,119 @@ class Customer extends ApiResource
     {
         $url = $this->instanceUrl() . '/discount';
         list($response, $opts) = $this->_request('delete', $url);
-        $this->refreshFrom(array('discount' => null), $opts, true);
+        $this->refreshFrom(['discount' => null], $opts, true);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer on which to create the source.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function createSource($id, $params = null, $opts = null)
+    {
+        return self::_createNestedResource($id, static::PATH_SOURCES, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer to which the source belongs.
+     * @param string|null $sourceId The ID of the source to retrieve.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function retrieveSource($id, $sourceId, $params = null, $opts = null)
+    {
+        return self::_retrieveNestedResource($id, static::PATH_SOURCES, $sourceId, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer to which the source belongs.
+     * @param string|null $sourceId The ID of the source to update.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function updateSource($id, $sourceId, $params = null, $opts = null)
+    {
+        return self::_updateNestedResource($id, static::PATH_SOURCES, $sourceId, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer to which the source belongs.
+     * @param string|null $sourceId The ID of the source to delete.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function deleteSource($id, $sourceId, $params = null, $opts = null)
+    {
+        return self::_deleteNestedResource($id, static::PATH_SOURCES, $sourceId, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer on which to retrieve the sources.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return Collection The list of sources.
+     */
+    public static function allSources($id, $params = null, $opts = null)
+    {
+        return self::_allNestedResources($id, static::PATH_SOURCES, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer on which to create the tax id.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function createTaxId($id, $params = null, $opts = null)
+    {
+        return self::_createNestedResource($id, static::PATH_TAX_IDS, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer to which the tax id belongs.
+     * @param string|null $taxIdId The ID of the tax id to retrieve.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function retrieveTaxId($id, $taxIdId, $params = null, $opts = null)
+    {
+        return self::_retrieveNestedResource($id, static::PATH_TAX_IDS, $taxIdId, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer to which the tax id belongs.
+     * @param string|null $taxIdId The ID of the tax id to delete.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return ApiResource
+     */
+    public static function deleteTaxId($id, $taxIdId, $params = null, $opts = null)
+    {
+        return self::_deleteNestedResource($id, static::PATH_TAX_IDS, $taxIdId, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the customer on which to retrieve the tax ids.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return Collection The list of tax ids.
+     */
+    public static function allTaxIds($id, $params = null, $opts = null)
+    {
+        return self::_allNestedResources($id, static::PATH_TAX_IDS, $params, $opts);
     }
 }
