@@ -8,7 +8,7 @@ namespace Stripe;
  * @property string $id
  * @property string $object
  * @property int $amount
- * @property mixed $balance_transactions
+ * @property BalanceTransaction[] $balance_transactions
  * @property string $charge
  * @property int $created
  * @property string $currency
@@ -16,7 +16,7 @@ namespace Stripe;
  * @property mixed $evidence_details
  * @property bool $is_charge_refundable
  * @property bool $livemode
- * @property mixed $metadata
+ * @property StripeObject $metadata
  * @property string $reason
  * @property string $status
  *
@@ -24,50 +24,44 @@ namespace Stripe;
  */
 class Dispute extends ApiResource
 {
-    /**
-     * @param array|string $id The ID of the dispute to retrieve, or an options
-     *     array containing an `id` key.
-     * @param array|string|null $options
-     *
-     * @return Dispute
-     */
-    public static function retrieve($id, $options = null)
-    {
-        return self::_retrieve($id, $options);
-    }
+
+    const OBJECT_NAME = "dispute";
+
+    use ApiOperations\All;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
 
     /**
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return array An array of Disputes.
+     * Possible string representations of dispute reasons.
+     * @link https://stripe.com/docs/api#dispute_object
      */
-    public static function all($params = null, $options = null)
-    {
-        return self::_all($params, $options);
-    }
+    const REASON_BANK_CANNOT_PROCESS       = 'bank_cannot_process';
+    const REASON_CHECK_RETURNED            = 'check_returned';
+    const REASON_CREDIT_NOT_PROCESSED      = 'credit_not_processed';
+    const REASON_CUSTOMER_INITIATED        = 'customer_initiated';
+    const REASON_DEBIT_NOT_AUTHORIZED      = 'debit_not_authorized';
+    const REASON_DUPLICATE                 = 'duplicate';
+    const REASON_FRAUDULENT                = 'fraudulent';
+    const REASON_GENERAL                   = 'general';
+    const REASON_INCORRECT_ACCOUNT_DETAILS = 'incorrect_account_details';
+    const REASON_INSUFFICIENT_FUNDS        = 'insufficient_funds';
+    const REASON_PRODUCT_NOT_RECEIVED      = 'product_not_received';
+    const REASON_PRODUCT_UNACCEPTABLE      = 'product_unacceptable';
+    const REASON_SUBSCRIPTION_CANCELED     = 'subscription_canceled';
+    const REASON_UNRECOGNIZED              = 'unrecognized';
 
     /**
-     * @param string $id The ID of the dispute to update.
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Dispute The updated dispute.
+     * Possible string representations of dispute statuses.
+     * @link https://stripe.com/docs/api#dispute_object
      */
-    public static function update($id, $params = null, $options = null)
-    {
-        return self::_update($id, $params, $options);
-    }
-
-    /**
-     * @param array|string|null $options
-     *
-     * @return Dispute The saved charge.
-     */
-    public function save($options = null)
-    {
-        return $this->_save($options);
-    }
+    const STATUS_CHARGE_REFUNDED        = 'charge_refunded';
+    const STATUS_LOST                   = 'lost';
+    const STATUS_NEEDS_RESPONSE         = 'needs_response';
+    const STATUS_UNDER_REVIEW           = 'under_review';
+    const STATUS_WARNING_CLOSED         = 'warning_closed';
+    const STATUS_WARNING_NEEDS_RESPONSE = 'warning_needs_response';
+    const STATUS_WARNING_UNDER_REVIEW   = 'warning_under_review';
+    const STATUS_WON                    = 'won';
 
     /**
      * @param array|string|null $options
