@@ -14,9 +14,7 @@ abstract class OAuth
      */
     public static function authorizeUrl($params = null, $opts = null)
     {
-        if (!$params) {
-            $params = array();
-        }
+        $params = $params ?: [];
 
         $base = ($opts && array_key_exists('connect_base', $opts)) ? $opts['connect_base'] : Stripe::$connectBase;
 
@@ -24,7 +22,7 @@ abstract class OAuth
         if (!array_key_exists('response_type', $params)) {
             $params['response_type'] = 'code';
         }
-        $query = Util\Util::urlEncode($params);
+        $query = Util\Util::encodeParameters($params);
 
         return $base . '/oauth/authorize?' . $query;
     }
@@ -61,10 +59,7 @@ abstract class OAuth
      */
     public static function deauthorize($params = null, $opts = null)
     {
-        if (!$params) {
-            $params = array();
-        }
-
+        $params = $params ?: [];
         $base = ($opts && array_key_exists('connect_base', $opts)) ? $opts['connect_base'] : Stripe::$connectBase;
         $requestor = new ApiRequestor(null, $base);
         $params['client_id'] = self::_getClientId($params);
