@@ -8,18 +8,19 @@ namespace Stripe;
  * @property string $id
  * @property string $object
  * @property int $amount
+ * @property int $arrival_date
+ * @property bool $automatic
  * @property string $balance_transaction
- * @property string $cancellation_balance_transaction
  * @property int $created
  * @property string $currency
- * @property int $arrival_date
+ * @property string $description
  * @property string $destination
+ * @property string $failure_balance_transaction
  * @property string $failure_code
  * @property string $failure_message
  * @property bool $livemode
- * @property mixed $metadata
+ * @property StripeObject $metadata
  * @property string $method
- * @property string $recipient
  * @property string $source_type
  * @property string $statement_descriptor
  * @property string $status
@@ -29,51 +30,55 @@ namespace Stripe;
  */
 class Payout extends ApiResource
 {
-    /**
-     * @param array|string $id The ID of the payout to retrieve, or an options
-     *     array containing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return Payout
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
+
+    const OBJECT_NAME = "payout";
+
+    use ApiOperations\All;
+    use ApiOperations\Create;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
 
     /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of Payouts
+     * Types of payout failure codes.
+     * @link https://stripe.com/docs/api#payout_failures
      */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
-    }
+    const FAILURE_ACCOUNT_CLOSED                = 'account_closed';
+    const FAILURE_ACCOUNT_FROZEN                = 'account_frozen';
+    const FAILURE_BANK_ACCOUNT_RESTRICTED       = 'bank_account_restricted';
+    const FAILURE_BANK_OWNERSHIP_CHANGED        = 'bank_ownership_changed';
+    const FAILURE_COULD_NOT_PROCESS             = 'could_not_process';
+    const FAILURE_DEBIT_NOT_AUTHORIZED          = 'debit_not_authorized';
+    const FAILURE_DECLINED                      = 'declined';
+    const FAILURE_INCORRECT_ACCOUNT_HOLDER_NAME = 'incorrect_account_holder_name';
+    const FAILURE_INSUFFICIENT_FUNDS            = 'insufficient_funds';
+    const FAILURE_INVALID_ACCOUNT_NUMBER        = 'invalid_account_number';
+    const FAILURE_INVALID_CURRENCY              = 'invalid_currency';
+    const FAILURE_NO_ACCOUNT                    = 'no_account';
+    const FAILURE_UNSUPPORTED_CARD              = 'unsupported_card';
 
     /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Payout The created payout.
+     * Possible string representations of the payout methods.
+     * @link https://stripe.com/docs/api/payouts/object#payout_object-method
      */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
+    const METHOD_STANDARD = 'standard';
+    const METHOD_INSTANT  = 'instant';
 
     /**
-     * @param string $id The ID of the payout to update.
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Payout The updated payout.
+     * Possible string representations of the status of the payout.
+     * @link https://stripe.com/docs/api/payouts/object#payout_object-status
      */
-    public static function update($id, $params = null, $options = null)
-    {
-        return self::_update($id, $params, $options);
-    }
+    const STATUS_CANCELED   = 'canceled';
+    const STATUS_IN_TRANSIT = 'in_transit';
+    const STATUS_FAILED     = 'failed';
+    const STATUS_PAID       = 'paid';
+    const STATUS_PENDING    = 'pending';
+
+    /**
+     * Possible string representations of the type of payout.
+     * @link https://stripe.com/docs/api/payouts/object#payout_object-type
+     */
+    const TYPE_BANK_ACCOUNT = 'bank_account';
+    const TYPE_CARD         = 'card';
 
     /**
      * @return Payout The canceled payout.
@@ -84,15 +89,5 @@ class Payout extends ApiResource
         list($response, $opts) = $this->_request('post', $url);
         $this->refreshFrom($response, $opts);
         return $this;
-    }
-
-    /**
-     * @param array|string|null $opts
-     *
-     * @return Payout The saved payout.
-     */
-    public function save($opts = null)
-    {
-        return $this->_save($opts);
     }
 }
