@@ -5,74 +5,43 @@ namespace Stripe;
 /**
  * Class Product
  *
+ * @property string $id
+ * @property string $object
+ * @property bool $active
+ * @property string[] $attributes
+ * @property string $caption
+ * @property int $created
+ * @property string[] $deactivate_on
+ * @property string $description
+ * @property string[] $images
+ * @property bool $livemode
+ * @property StripeObject $metadata
+ * @property string $name
+ * @property mixed $package_dimensions
+ * @property bool $shippable
+ * @property string $statement_descriptor
+ * @property string $type
+ * @property string $unit_label
+ * @property int $updated
+ * @property string $url
+ *
  * @package Stripe
  */
 class Product extends ApiResource
 {
-    /**
-     * @param array|string $id The ID of the product to retrieve, or an options
-     *     array contianing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return Product
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
+
+    const OBJECT_NAME = "product";
+
+    use ApiOperations\All;
+    use ApiOperations\Create;
+    use ApiOperations\Delete;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
 
     /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Product The created Product.
+     * Possible string representations of the type of product.
+     * @link https://stripe.com/docs/api/service_products/object#service_product_object-type
      */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
-
-    /**
-     * @param string $id The ID of the product to update.
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Product The updated product.
-     */
-    public static function update($id, $params = null, $options = null)
-    {
-        return self::_update($id, $params, $options);
-    }
-
-    /**
-     * @param array|string|null $opts
-     *
-     * @return Product The saved Product.
-     */
-    public function save($opts = null)
-    {
-        return $this->_save($opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of Products
-     */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Product The deleted product.
-     */
-    public function delete($params = null, $opts = null)
-    {
-        return $this->_delete($params, $opts);
-    }
+    const TYPE_GOOD    = 'good';
+    const TYPE_SERVICE = 'service';
 }
