@@ -20,7 +20,7 @@ namespace Stripe;
  * @property string $failure_code
  * @property string $failure_message
  * @property bool $livemode
- * @property mixed $metadata
+ * @property StripeObject $metadata
  * @property string $method
  * @property string $recipient
  * @property mixed $reversals
@@ -35,4 +35,5 @@ namespace Stripe;
 class RecipientTransfer extends ApiResource
 {
 
+    const OBJECT_NAME = "recipient_transfer";
 }
