@@ -22,7 +22,9 @@ abstract class SingletonApiResource extends ApiResource
      */
     public static function classUrl()
     {
-        $base = static::className();
+        // Replace dots with slashes for namespaced resources, e.g. if the object's name is
+        // "foo.bar", then its URL will be "/v1/foo/bar".
+        $base = str_replace('.', '/', static::OBJECT_NAME);
         return "/v1/${base}";
     }
 
