@@ -5,73 +5,84 @@ namespace Stripe;
 /**
  * Class Source
  *
+ * @property string $id
+ * @property string $object
+ * @property mixed $ach_credit_transfer
+ * @property mixed $ach_debit
+ * @property mixed $alipay
+ * @property int $amount
+ * @property mixed $bancontact
+ * @property mixed $card
+ * @property mixed $card_present
+ * @property string $client_secret
+ * @property mixed $code_verification
+ * @property int $created
+ * @property string $currency
+ * @property string $customer
+ * @property mixed $eps
+ * @property string $flow
+ * @property mixed $giropay
+ * @property mixed $ideal
+ * @property bool $livemode
+ * @property StripeObject $metadata
+ * @property mixed $multibanco
+ * @property mixed $owner
+ * @property mixed $p24
+ * @property mixed $receiver
+ * @property mixed $redirect
+ * @property mixed $sepa_debit
+ * @property mixed $sofort
+ * @property string $statement_descriptor
+ * @property string $status
+ * @property mixed $three_d_secure
+ * @property string $type
+ * @property string $usage
+ * @property mixed $wechat
+
  * @package Stripe
  */
 class Source extends ApiResource
 {
-    /**
-     * @param array|string $id The ID of the source to retrieve, or an options
-     *     array containing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return Source
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
+
+    const OBJECT_NAME = "source";
+
+    use ApiOperations\Create;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
 
     /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of Sources
+     * Possible string representations of source flows.
+     * @link https://stripe.com/docs/api#source_object-flow
      */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
-    }
+    const FLOW_REDIRECT          = 'redirect';
+    const FLOW_RECEIVER          = 'receiver';
+    const FLOW_CODE_VERIFICATION = 'code_verification';
+    const FLOW_NONE              = 'none';
 
     /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Source The created Source.
+     * Possible string representations of source statuses.
+     * @link https://stripe.com/docs/api#source_object-status
      */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
+    const STATUS_CANCELED   = 'canceled';
+    const STATUS_CHARGEABLE = 'chargeable';
+    const STATUS_CONSUMED   = 'consumed';
+    const STATUS_FAILED     = 'failed';
+    const STATUS_PENDING    = 'pending';
 
     /**
-     * @param string $id The ID of the source to update.
+     * Possible string representations of source usage.
+     * @link https://stripe.com/docs/api#source_object-usage
+     */
+    const USAGE_REUSABLE   = 'reusable';
+    const USAGE_SINGLE_USE = 'single_use';
+
+    /**
      * @param array|null $params
      * @param array|string|null $options
      *
-     * @return Source The updated source.
+     * @return Source The detached source.
      */
-    public static function update($id, $params = null, $options = null)
-    {
-        return self::_update($id, $params, $options);
-    }
-
-    /**
-     * @param array|string|null $opts
-     *
-     * @return Source The saved source.
-     */
-    public function save($opts = null)
-    {
-        return $this->_save($opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Source The deleted source.
-     */
-    public function delete($params = null, $options = null)
+    public function detach($params = null, $options = null)
     {
         self::_validateParams($params);
 
@@ -93,9 +104,8 @@ class Source extends ApiResource
             $this->refreshFrom($response, $opts);
             return $this;
         } else {
-            $message = "Source objects cannot be deleted, they can only be "
-               . "detached from customer objects. This source object does not "
-               . "appear to be currently attached to a customer object.";
+            $message = "This source object does not appear to be currently attached "
+               . "to a customer object.";
             throw new Error\Api($message);
         }
     }
@@ -104,7 +114,35 @@ class Source extends ApiResource
      * @param array|null $params
      * @param array|string|null $options
      *
-     * @return BankAccount The verified bank account.
+     * @return Source The detached source.
+     *
+     * @deprecated Use the `detach` method instead.
+     */
+    public function delete($params = null, $options = null)
+    {
+        $this->detach($params, $options);
+    }
+
+    /**
+     * @param array|null $params
+     * @param array|string|null $options
+     *
+     * @return Collection The list of source transactions.
+     */
+    public function sourceTransactions($params = null, $options = null)
+    {
+        $url = $this->instanceUrl() . '/source_transactions';
+        list($response, $opts) = $this->_request('get', $url, $params, $options);
+        $obj = Util\Util::convertToStripeObject($response, $opts);
+        $obj->setLastResponse($response);
+        return $obj;
+    }
+
+    /**
+     * @param array|null $params
+     * @param array|string|null $options
+     *
+     * @return Source The verified source.
      */
     public function verify($params = null, $options = null)
     {
