@@ -5,85 +5,41 @@ namespace Stripe;
 /**
  * Class SubscriptionItem
  *
+ * @property string $id
+ * @property string $object
+ * @property mixed $billing_thresholds
+ * @property int $created
+ * @property StripeObject $metadata
+ * @property Plan $plan
+ * @property int $quantity
+ * @property string $subscription
+ * @property array $tax_rates
+ *
  * @package Stripe
  */
 class SubscriptionItem extends ApiResource
 {
-    /**
-     * This is a special case because the subscription items endpoint has an
-     *    underscore in it. The parent `className` function strips underscores.
-     *
-     * @return string The name of the class.
-     */
-    public static function className()
-    {
-        return 'subscription_item';
-    }
+
+    const OBJECT_NAME = "subscription_item";
+
+    use ApiOperations\All;
+    use ApiOperations\Create;
+    use ApiOperations\Delete;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
 
     /**
-     * @param array|string $id The ID of the subscription item to retrieve, or
-     *     an options array containing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return SubscriptionItem
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of SubscriptionItems
-     */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return SubscriptionItem The created subscription item.
-     */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
-
-    /**
-     * @param string $id The ID of the subscription item to update.
      * @param array|null $params
      * @param array|string|null $options
      *
-     * @return SubscriptionItem The updated subscription item.
+     * @return Collection The list of source transactions.
      */
-    public static function update($id, $params = null, $options = null)
+    public function usageRecordSummaries($params = null, $options = null)
     {
-        return self::_update($id, $params, $options);
-    }
-
-    /**
-     * @param array|string|null $opts
-     *
-     * @return SubscriptionItem The saved subscription item.
-     */
-    public function save($opts = null)
-    {
-        return $this->_save($opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return SubscriptionItem The deleted subscription item.
-     */
-    public function delete($params = null, $opts = null)
-    {
-        return $this->_delete($params, $opts);
+        $url = $this->instanceUrl() . '/usage_record_summaries';
+        list($response, $opts) = $this->_request('get', $url, $params, $options);
+        $obj = Util\Util::convertToStripeObject($response, $opts);
+        $obj->setLastResponse($response);
+        return $obj;
     }
 }
