@@ -12,64 +12,40 @@ namespace Stripe;
  * @property string $balance_transaction
  * @property int $created
  * @property string $currency
- * @property int $date
- * @property mixed $destination
- * @property mixed $destination_payment
+ * @property string $description
+ * @property string $destination
+ * @property string $destination_payment
  * @property bool $livemode
- * @property mixed $metadata
- * @property mixed $reversals
+ * @property StripeObject $metadata
+ * @property Collection $reversals
  * @property bool $reversed
- * @property mixed $source_transaction
+ * @property string $source_transaction
+ * @property string $source_type
+ * @property string $transfer_group
  *
  * @package Stripe
  */
 class Transfer extends ApiResource
 {
-    /**
-     * @param array|string $id The ID of the transfer to retrieve, or an
-     *     options array containing an `id` key.
-     * @param array|string|null $opts
-     *
-     * @return Transfer
-     */
-    public static function retrieve($id, $opts = null)
-    {
-        return self::_retrieve($id, $opts);
-    }
+
+    const OBJECT_NAME = "transfer";
+
+    use ApiOperations\All;
+    use ApiOperations\Create;
+    use ApiOperations\NestedResource;
+    use ApiOperations\Retrieve;
+    use ApiOperations\Update;
+
+    const PATH_REVERSALS = '/reversals';
 
     /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Collection of Transfers
+     * Possible string representations of the source type of the transfer.
+     * @link https://stripe.com/docs/api/transfers/object#transfer_object-source_type
      */
-    public static function all($params = null, $opts = null)
-    {
-        return self::_all($params, $opts);
-    }
-
-    /**
-     * @param array|null $params
-     * @param array|string|null $opts
-     *
-     * @return Transfer The created transfer.
-     */
-    public static function create($params = null, $opts = null)
-    {
-        return self::_create($params, $opts);
-    }
-
-    /**
-     * @param string $id The ID of the transfer to update.
-     * @param array|null $params
-     * @param array|string|null $options
-     *
-     * @return Transfer The updated transfer.
-     */
-    public static function update($id, $params = null, $options = null)
-    {
-        return self::_update($id, $params, $options);
-    }
+    const SOURCE_TYPE_ALIPAY_ACCOUNT = 'alipay_account';
+    const SOURCE_TYPE_BANK_ACCOUNT   = 'bank_account';
+    const SOURCE_TYPE_CARD           = 'card';
+    const SOURCE_TYPE_FINANCING      = 'financing';
 
     /**
      * @return TransferReversal The created transfer reversal.
@@ -94,12 +70,52 @@ class Transfer extends ApiResource
     }
 
     /**
+     * @param string|null $id The ID of the transfer on which to create the reversal.
+     * @param array|null $params
      * @param array|string|null $opts
      *
-     * @return Transfer The saved transfer.
+     * @return TransferReversal
      */
-    public function save($opts = null)
+    public static function createReversal($id, $params = null, $opts = null)
     {
-        return $this->_save($opts);
+        return self::_createNestedResource($id, static::PATH_REVERSALS, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the transfer to which the reversal belongs.
+     * @param array|null $reversalId The ID of the reversal to retrieve.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return TransferReversal
+     */
+    public static function retrieveReversal($id, $reversalId, $params = null, $opts = null)
+    {
+        return self::_retrieveNestedResource($id, static::PATH_REVERSALS, $reversalId, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the transfer to which the reversal belongs.
+     * @param array|null $reversalId The ID of the reversal to update.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return TransferReversal
+     */
+    public static function updateReversal($id, $reversalId, $params = null, $opts = null)
+    {
+        return self::_updateNestedResource($id, static::PATH_REVERSALS, $reversalId, $params, $opts);
+    }
+
+    /**
+     * @param string|null $id The ID of the transfer on which to retrieve the reversals.
+     * @param array|null $params
+     * @param array|string|null $opts
+     *
+     * @return Collection The list of reversals.
+     */
+    public static function allReversals($id, $params = null, $opts = null)
+    {
+        return self::_allNestedResources($id, static::PATH_REVERSALS, $params, $opts);
     }
 }
