@@ -20,10 +20,12 @@ abstract class Webhook
      *  timestamp and the current time
      * @return \Stripe\Event the Event instance
      * @throws \UnexpectedValueException if the payload is not valid JSON,
-     *  \Stripe\SignatureVerification if the verification fails.
+     * @throws \Stripe\Error\SignatureVerification if the verification fails.
      */
     public static function constructEvent($payload, $sigHeader, $secret, $tolerance = self::DEFAULT_TOLERANCE)
     {
+        WebhookSignature::verifyHeader($payload, $sigHeader, $secret, $tolerance);
+
         $data = json_decode($payload, true);
         $jsonError = json_last_error();
         if ($data === null && $jsonError !== JSON_ERROR_NONE) {
@@ -31,9 +33,7 @@ abstract class Webhook
               . "(json_last_error() was $jsonError)";
             throw new \UnexpectedValueException($msg);
         }
-        $event = Event::constructFrom($data, null);
-
-        WebhookSignature::verifyHeader($payload, $sigHeader, $secret, $tolerance);
+        $event = Event::constructFrom($data);
 
         return $event;
     }
